+++ flutter/lib/src/sentry_flutter.dart
@@ -9,6 +9,7 @@ import '../sentry_flutter.dart';
 import 'event_processor/android_platform_exception_event_processor.dart';
 import 'event_processor/flutter_exception_event_processor.dart';
 import 'integrations/frame_tracking_integration.dart';
+import 'event_processor/platform_exception_event_processor.dart';
 import 'integrations/screenshot_integration.dart';
 import 'native_scope_observer.dart';
 import 'renderer/renderer.dart';
@@ -113,6 +114,8 @@ mixin SentryFlutter {
       options
           .addEventProcessor(AndroidPlatformExceptionEventProcessor(options));
     }
+
+    options.addEventProcessor(PlatformExceptionEventProcessor());
 
     _setSdk(options);
   }